Fix pointer/int warnings so this compiles on amd64.  The driver wants to store an int in the bio->bio_driver1 (a void *).  It is big enough, but you have to match the int sizes first before doing the cast.
Glanced at by:  scottl

+++ sys/dev/ips/ips_commands.c
@@ -97,7 +97,7 @@ static void ips_io_request_callback(void *cmdptr, bus_dma_segment_t *segments,in
 	}
 	command_struct = (ips_io_cmd *)command->command_buffer;
 	command_struct->id = command->id;
-	command_struct->drivenum = (uint32_t)iobuf->bio_driver1;
+	command_struct->drivenum = (uintptr_t)iobuf->bio_driver1;
 	if(segnum != 1){
 		if(ips_read_request(iobuf))
 			cmdtype = IPS_SG_READ_CMD;
